tests/requirements/base.txt
# requirements from setup.py
djangorestframework
djangocms-text
django-filer
setuptools

# other requirements
coverage
tox
